fix(#0): renaming file and fix usage - logging was renamed as logging_a now better as logging_amqtt with fix inside pyproject.toml - renamed codecs => codecs_a => codecs_amqtt
