+++ amqtt/mqtt/connect.py
@@ -2,7 +2,7 @@ from asyncio import StreamReader
 from typing import Self
 
 from amqtt.adapters import ReaderAdapter
-from amqtt.codecs_a import (
+from amqtt.codecs_amqtt import (
     bytes_to_int,
     decode_data_with_length,
     decode_string,
