+++ amqtt/mqtt/publish.py
@@ -2,7 +2,7 @@ import asyncio
 from typing import Self
 
 from amqtt.adapters import ReaderAdapter
-from amqtt.codecs_a import decode_packet_id, decode_string, encode_string, int_to_bytes
+from amqtt.codecs_amqtt import decode_packet_id, decode_string, encode_string, int_to_bytes
 from amqtt.errors import AMQTTError, MQTTError
 from amqtt.mqtt.packet import PUBLISH, MQTTFixedHeader, MQTTPacket, MQTTPayload, MQTTVariableHeader
