+++ amqtt/plugins/logging_amqtt.py
